Move utils/functions code to utils/index
